+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.3.0 / unreleased
 
 * [FEATURE] New optional parameter open(..., extraData).
+* [FEATURE] New option `autoTrigger: true` can be set to `false` to prevent opening on browser's `contextmenu` event.
 * [CHANGE] `setEntry()` and `replaceMenu()` now allow to define titles with HTML markup.
 
 # 1.2.4 / 2013-12-25